+++ src/base/Str.h
@@ -194,6 +194,8 @@ WStr Replace(WStr s, WStr toReplace, WStr replaceWith);
 namespace url {
 
 TempStr DecodeTemp(Str url);
+TempStr EncodeTemp(Str s);
+TempStr EncodeMayTruncateTemp(Str s, int maxEncodedLen, bool* didTruncateOut = nullptr);
 bool IsAbsolute(Str url);
 TempStr GetFullPathTemp(Str url);
 TempStr GetFileNameTemp(Str url);